security: Require team session tokens on AI routes
Move the four /api/ai/* endpoints into a dedicated aiRoutes.js module and
reject any request that does not carry a valid team session token for an
existing team, so anonymous network callers can no longer drive the
internal LLM. The client now sends the token it already holds from
login/restore/invite at every AI call site, keeping the change invisible
to users. Bump VERSION to 27.8 (security, no changelog entry).
