+++ VERSION
@@ -1,1 +1,1 @@
-27.7
+27.8
