+++ __tests__/routeHardening.test.ts
@@ -344,79 +344,4 @@ describe('route hardening', () => {
     expect(await actionResponse.json()).toEqual({ success: true });
   });
 
-  it('truncates release analysis prompt fields before calling the AI service', async () => {
-    const app = express();
-    const generateReleaseAnalysis = vi.fn(async () => 'analysis');
-    app.use(express.json({ limit: '2mb' }));
-    registerSuperAdminRoutes({
-      app,
-      io: { emit: vi.fn() },
-      dataStore: {},
-      tokenService: {
-        validateSuperAdminAuth: vi.fn(),
-        validateSuperAdminToken: vi.fn(),
-        createSuperAdminToken: vi.fn()
-      },
-      mailerService: {},
-      logService: { addServerLog: vi.fn(), getServerLogs: vi.fn(() => []) },
-      escapeHtml: (value: string) => value,
-      superAdminPassword: 'secret',
-      sessionCache: { clear: vi.fn() },
-      backupService: {},
-      aiService: { generateReleaseAnalysis }
-    });
-
-    const retrospectives = Array.from({ length: 50 }, (_, index) => ({ id: `retro-${index}` }));
-    const response = await request(app, '/api/ai/generate-release-analysis', {
-      method: 'POST',
-      headers: { 'content-type': 'application/json' },
-      body: JSON.stringify({
-        retrospectives,
-        customPrompt: 'x'.repeat(5000),
-        additionalInstructions: 'y'.repeat(5000)
-      })
-    });
-
-    expect(response.status).toBe(200);
-    expect(await response.json()).toEqual({ analysis: 'analysis' });
-    expect(generateReleaseAnalysis).toHaveBeenCalledWith(expect.objectContaining({
-      retrospectives,
-      customPrompt: 'x'.repeat(4000),
-      additionalInstructions: 'y'.repeat(4000)
-    }));
-  });
-
-  it('rejects release analysis requests that would silently omit selected retrospectives', async () => {
-    const app = express();
-    const generateReleaseAnalysis = vi.fn(async () => 'analysis');
-    app.use(express.json({ limit: '2mb' }));
-    registerSuperAdminRoutes({
-      app,
-      io: { emit: vi.fn() },
-      dataStore: {},
-      tokenService: {
-        validateSuperAdminAuth: vi.fn(),
-        validateSuperAdminToken: vi.fn(),
-        createSuperAdminToken: vi.fn()
-      },
-      mailerService: {},
-      logService: { addServerLog: vi.fn(), getServerLogs: vi.fn(() => []) },
-      escapeHtml: (value: string) => value,
-      superAdminPassword: 'secret',
-      sessionCache: { clear: vi.fn() },
-      backupService: {},
-      aiService: { generateReleaseAnalysis }
-    });
-
-    const response = await request(app, '/api/ai/generate-release-analysis', {
-      method: 'POST',
-      headers: { 'content-type': 'application/json' },
-      body: JSON.stringify({ retrospectives: Array.from({ length: 51 }, (_, index) => ({ id: `retro-${index}` })) })
-    });
-
-    expect(response.status).toBe(400);
-    expect(await response.json()).toEqual({ error: 'too_many_retrospectives', maxRetrospectives: 50 });
-    expect(generateReleaseAnalysis).not.toHaveBeenCalled();
-  });
-
 });
